refactor(hub-common): computeLinks calls computeItemLinks
affects: @esri/hub-common
