+++ packages/common/src/initiative-templates/_internal/computeLinks.ts
@@ -1,11 +1,7 @@
-import { UserSession } from "@esri/arcgis-rest-auth";
 import { IRequestOptions } from "@esri/arcgis-rest-request";
 import { IItem } from "@esri/arcgis-rest-types";
-import { getHubRelativeUrl } from "../../content/_internal/internalContentUtils";
-import { getRelativeWorkspaceUrl, IHubEntityLinks } from "../../core";
-import { getItemIdentifier } from "../../items";
-import { getItemThumbnailUrl } from "../../resources";
-import { getItemHomeUrl } from "../../urls";
+import { IHubEntityLinks } from "../../core";
+import { computeItemLinks } from "../../core/_internal/computeItemLinks";
 
 /**
  * Compute the links that get appended to a Hub Initiative Template
@@ -19,20 +15,5 @@ export function computeLinks(
   item: IItem,
   requestOptions: IRequestOptions
 ): IHubEntityLinks {
-  let token: string;
-  if (requestOptions.authentication) {
-    const session: UserSession = requestOptions.authentication as UserSession;
-    token = session.token;
-  }
-
-  return {
-    self: getItemHomeUrl(item.id, requestOptions),
-    siteRelative: getHubRelativeUrl(item.type, getItemIdentifier(item)),
-    siteRelativeEntityType: getHubRelativeUrl(item.type),
-    workspaceRelative: getRelativeWorkspaceUrl(
-      item.type,
-      getItemIdentifier(item)
-    ),
-    thumbnail: getItemThumbnailUrl(item, requestOptions, token),
-  };
+  return computeItemLinks(item, requestOptions);
 }
