+++ packages/common/src/projects/_internal/computeLinks.ts
@@ -1,12 +1,7 @@
-import { IItem } from "@esri/arcgis-rest-types";
+import { computeItemLinks } from "../../core/_internal/computeItemLinks";
 import { IRequestOptions } from "@esri/arcgis-rest-request";
-import { UserSession } from "@esri/arcgis-rest-auth";
-import { getItemHomeUrl } from "../../urls";
-import { IHubEntityLinks } from "../../core/types";
-import { getItemIdentifier } from "../../items";
-import { getHubRelativeUrl } from "../../content/_internal/internalContentUtils";
-import { getRelativeWorkspaceUrl } from "../../core/getRelativeWorkspaceUrl";
-import { getItemThumbnailUrl } from "../../resources/get-item-thumbnail-url";
+import { IItem } from "@esri/arcgis-rest-types";
+import { IHubEntityLinks } from "../../core";
 
 /**
  * Compute the links that get appended to a Hub Project
@@ -14,25 +9,11 @@ import { getItemThumbnailUrl } from "../../resources/get-item-thumbnail-url";
  *
  * @param item
  * @param requestOptions
+ * @returns
  */
 export function computeLinks(
   item: IItem,
   requestOptions: IRequestOptions
 ): IHubEntityLinks {
-  let token: string;
-  if (requestOptions.authentication) {
-    const session: UserSession = requestOptions.authentication as UserSession;
-    token = session.token;
-  }
-
-  return {
-    self: getItemHomeUrl(item.id, requestOptions),
-    siteRelative: getHubRelativeUrl(item.type, getItemIdentifier(item)),
-    siteRelativeEntityType: getHubRelativeUrl(item.type),
-    workspaceRelative: getRelativeWorkspaceUrl(
-      item.type,
-      getItemIdentifier(item)
-    ),
-    thumbnail: getItemThumbnailUrl(item, requestOptions, token),
-  };
+  return computeItemLinks(item, requestOptions);
 }
