+++ packages/common/src/sites/_internal/computeLinks.ts
@@ -1,11 +1,8 @@
 import { IItem } from "@esri/arcgis-rest-types";
 import { IRequestOptions } from "@esri/arcgis-rest-request";
-import { UserSession } from "@esri/arcgis-rest-auth";
 import { IHubEntityLinks } from "../../core/types";
-import { getItemIdentifier } from "../../items";
-import { getRelativeWorkspaceUrl } from "../../core/getRelativeWorkspaceUrl";
-import { getItemThumbnailUrl } from "../../resources/get-item-thumbnail-url";
 import { getHubRelativeUrl } from "../../content/_internal/internalContentUtils";
+import { computeItemLinks } from "../../core/_internal/computeItemLinks";
 
 /**
  * Compute the links that get appended to a Hub Site
@@ -18,18 +15,12 @@ export function computeLinks(
   item: IItem,
   requestOptions: IRequestOptions
 ): IHubEntityLinks {
-  let token: string;
-  if (requestOptions.authentication) {
-    const session: UserSession = requestOptions.authentication as UserSession;
-    token = session.token;
-  }
-
+  const links = computeItemLinks(item, requestOptions);
   return {
+    ...links,
+    // for sites we use the site's url as the self link
     self: item.url,
-    siteRelative: getHubRelativeUrl(item.type, item.id, item.typeKeywords),
-    siteRelativeEntityType: getHubRelativeUrl(item.type),
+    // add the layout relative link
     layoutRelative: "/edit",
-    workspaceRelative: getRelativeWorkspaceUrl(item.type, item.id),
-    thumbnail: getItemThumbnailUrl(item, requestOptions, token),
   };
 }
